Refactor Project loading code
