+++ components/sections/projects/index.tsx
@@ -1,8 +1,8 @@
 import * as S from "./styles";
 import { CardRow } from "components/layout";
-import { PortalProject } from "lib/portal-types";
 import strings from "content/strings.json";
 import { Route } from "lib/utils";
+import { PortalProject } from "lib/airtable/project";
 
 const msg = strings.components.sections.projects;
